pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-analyst"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2,<2",
    "langchain-community>=0.4,<1",
    "langchain-classic>=1.0,<2",
    "langchain-openai>=1.1,<2",
    "langchain-text-splitters>=1.1,<2",
    "chromadb>=1.0,<2",
    "openai>=2.0,<3",
    "pymupdf>=1.20,<2",
    "tiktoken>=0.12,<1",
    "rank-bm25>=0.2,<1",
    "python-dotenv>=1.0,<2",
    "pydantic>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "nbstripout>=0.7",
]

[project.scripts]
rag-analyst = "rag_analyst.runner:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_analyst"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
